feat(spec-renderer): move spec-details package
