--- packages/portal/spec-details/CHANGELOG.md
@@ -1,200 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.5.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.5.1...@kong-ui-public/spec-details@0.5.2) (2023-02-02)
-
-
-### ♻️ Chores
-
-* **deps:** update dependency vue to ^3.2.47 ([#98](https://github.com/Kong/public-ui-components/issues/98)) ([eb28034](https://github.com/Kong/public-ui-components/commit/eb28034b5707b1000762038e7218c305b53c831c))
-
-
-
-
-
-## [0.5.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.5.0...@kong-ui-public/spec-details@0.5.1) (2023-02-02)
-
-
-### ♻️ Chores
-
-* **deps:** update dependency vue to ^3.2.46 ([#96](https://github.com/Kong/public-ui-components/issues/96)) ([cc1d5f1](https://github.com/Kong/public-ui-components/commit/cc1d5f16228d5ea6b81222c67aca8e7d33db9d4c))
-
-
-
-
-
-# [0.5.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.4.0...@kong-ui-public/spec-details@0.5.0) (2023-02-01)
-
-
-### ✨ Features
-
-* **swagger-ui:** add fire event on button presses TDX-2778 ([#81](https://github.com/Kong/public-ui-components/issues/81)) ([dab3101](https://github.com/Kong/public-ui-components/commit/dab3101001c9713f7e35860942a8c9f455dfcb6f))
-
-
-
-
-
-# [0.4.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.3.7...@kong-ui-public/spec-details@0.4.0) (2023-02-01)
-
-
-### ✨ Features
-
-* **spec-details:** add activeOperation property ([#92](https://github.com/Kong/public-ui-components/issues/92)) ([0350ee9](https://github.com/Kong/public-ui-components/commit/0350ee96b5bacf9f5728f267142c0b1958185f10))
-
-
-
-
-
-## [0.3.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.3.6...@kong-ui-public/spec-details@0.3.7) (2023-01-27)
-
-**Note:** Version bump only for package @kong-ui-public/spec-details
-
-
-
-
-
-## [0.3.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.3.5...@kong-ui-public/spec-details@0.3.6) (2023-01-27)
-
-**Note:** Version bump only for package @kong-ui-public/spec-details
-
-
-
-
-
-## [0.3.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.3.4...@kong-ui-public/spec-details@0.3.5) (2023-01-27)
-
-
-### ♻️ Chores
-
-* **stylelint:** add stylelint-order ([#61](https://github.com/Kong/public-ui-components/issues/61)) ([d28a9c4](https://github.com/Kong/public-ui-components/commit/d28a9c42819aa4e19cb294165c6458b973b33e83))
-
-
-
-
-
-## [0.3.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-details@0.3.3...@kong-ui-public/spec-details@0.3.4) (2023-01-27)
-
-
-### 🐛 Bug Fixes
-
-* **spec-details:** misc fixes ([#60](https://github.com/Kong/public-ui-components/issues/60)) ([8137337](https://github.com/Kong/public-ui-components/commit/813733757c6797b2bcc19f802b0d2dde31c52cbe))
-
-
-
-
-
-## 0.3.3 (2023-01-27)
-
-
-### ♻️ Chores
-
-* bump swagger-ui-kong-theme-universal, rename spec-renderer to spec-details ([#58](https://github.com/Kong/public-ui-components/issues/58)) ([5d9c2af](https://github.com/Kong/public-ui-components/commit/5d9c2af5f8184d7f353f5e5f2437fdaabe99a673))
-
-
-
-
-
-## [0.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.3.1...@kong-ui-public/spec-renderer@0.3.2) (2023-01-24)
-
-**Note:** Version bump only for package @kong-ui-public/spec-renderer
-
-
-
-
-
-## [0.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.3.0...@kong-ui-public/spec-renderer@0.3.1) (2023-01-24)
-
-**Note:** Version bump only for package @kong-ui-public/spec-renderer
-
-
-
-
-
-# [0.3.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.2.2...@kong-ui-public/spec-renderer@0.3.0) (2023-01-24)
-
-
-### ✨ Features
-
-* **portal:** create swagger-ui-web-component package ([#42](https://github.com/Kong/public-ui-components/issues/42)) ([5db0c93](https://github.com/Kong/public-ui-components/commit/5db0c9382b58ad76d7ad5e40aa4e3e2393b7d900))
-
-
-
-
-
-## [0.2.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.2.1...@kong-ui-public/spec-renderer@0.2.2) (2023-01-24)
-
-
-### ♻️ Chores
-
-* update package.json ([1b18450](https://github.com/Kong/public-ui-components/commit/1b184509aff84030da887951dfdadc6a3d52a986))
-
-
-
-
-
-## [0.2.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.2.0...@kong-ui-public/spec-renderer@0.2.1) (2023-01-24)
-
-
-### ♻️ Chores
-
-* update package.json repo ([afd1cb5](https://github.com/Kong/public-ui-components/commit/afd1cb579df4cf962a2920185086620cc32a91e1))
-
-
-
-
-
-# 0.2.0 (2023-01-24)
-
-
-### ✨ Features
-
-* **portal-spec-renderer:** create package ([#28](https://github.com/Kong/public-ui-components/issues/28)) ([d6514eb](https://github.com/Kong/public-ui-components/commit/d6514eb412f76026996f4e3082fccaed6b14b5cf))
-
-
-### 🐛 Bug Fixes
-
-* **deps:** update dependency [@kong](https://github.com/kong)/swagger-ui-web-component to ^1.0.17 ([#34](https://github.com/Kong/public-ui-components/issues/34)) ([b2049c9](https://github.com/Kong/public-ui-components/commit/b2049c9d1af14a4c498433add44148d78b6d2680))
-* **spec-renderer:** remove slimMode ([#35](https://github.com/Kong/public-ui-components/issues/35)) ([3e081e7](https://github.com/Kong/public-ui-components/commit/3e081e7db96a60416e12fb0e07a679759bc51cf6))
-
-
-### ♻️ Chores
-
-* change npm scope to [@kong-ui-public](https://github.com/kong-ui-public) ([#40](https://github.com/Kong/public-ui-components/issues/40)) ([c8c3c4f](https://github.com/Kong/public-ui-components/commit/c8c3c4f165fdd36269e1dd5bb31f47f73f7e9f4a))
-* **release:** publish [skip ci] ([a7b4182](https://github.com/Kong/public-ui-components/commit/a7b4182fdc856e6e31eeab07767017b4e060e3a9))
-* **release:** publish [skip ci] ([b14c686](https://github.com/Kong/public-ui-components/commit/b14c6862b589e645cb4dfd6cb4a40e7f9fd37cf2))
-* **release:** publish [skip ci] ([cf5eb7d](https://github.com/Kong/public-ui-components/commit/cf5eb7d6c6a6455e07132a0ab4e9209d9a759e48))
-
-
-
-
-
-## [0.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.1.1...@kong-ui-public/spec-renderer@0.1.2) (2023-01-23)
-
-
-### 🐛 Bug Fixes
-
-* **spec-renderer:** remove slimMode ([#35](https://github.com/Kong/public-ui-components/issues/35)) ([3e081e7](https://github.com/Kong/public-ui-components/commit/3e081e7db96a60416e12fb0e07a679759bc51cf6))
-
-
-
-
-
-## [0.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/spec-renderer@0.1.0...@kong-ui-public/spec-renderer@0.1.1) (2023-01-23)
-
-
-### 🐛 Bug Fixes
-
-* **deps:** update dependency [@kong](https://github.com/kong)/swagger-ui-web-component to ^1.0.17 ([#34](https://github.com/Kong/public-ui-components/issues/34)) ([b2049c9](https://github.com/Kong/public-ui-components/commit/b2049c9d1af14a4c498433add44148d78b6d2680))
-
-
-
-
-
-# 0.1.0 (2023-01-23)
-
-
-### ✨ Features
-
-* **portal-spec-renderer:** create package ([#28](https://github.com/Kong/public-ui-components/issues/28)) ([d6514eb](https://github.com/Kong/public-ui-components/commit/d6514eb412f76026996f4e3082fccaed6b14b5cf))
