--- packages/portal/spec-details/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "@kong-ui-public/spec-details",
-  "version": "0.5.2",
-  "type": "module",
-  "main": "./dist/spec-details.umd.js",
-  "module": "./dist/spec-details.es.js",
-  "types": "dist/types/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "exports": {
-    ".": {
-      "import": "./dist/spec-details.es.js",
-      "require": "./dist/spec-details.umd.js"
-    },
-    "./package.json": "./package.json",
-    "./dist/*": "./dist/*"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "dev": "cross-env USE_SANDBOX=true vite",
-    "build": "run-s typecheck build:package build:types",
-    "build:package": "vite build",
-    "build:types": "vue-tsc -p './tsconfig.build.json' --emitDeclarationOnly",
-    "build:visualize": "BUILD_VISUALIZER='portal/spec-details' vite build -m production",
-    "preview:package": "vite preview --port 4173",
-    "preview": "cross-env USE_SANDBOX=true PREVIEW_SANDBOX=true run-s build:package preview:package",
-    "lint": "eslint '**/*.{js,jsx,ts,tsx,vue}' --ignore-path '../../../.eslintignore'",
-    "lint:fix": "eslint '**/*.{js,jsx,ts,tsx,vue}' --ignore-path '../../../.eslintignore' --fix",
-    "stylelint": "stylelint --allow-empty-input './src/**/*.{css,scss,sass,less,styl,vue}'",
-    "stylelint:fix": "stylelint --allow-empty-input './src/**/*.{css,scss,sass,less,styl,vue}' --fix",
-    "typecheck": "vue-tsc -p './tsconfig.build.json' --noEmit",
-    "test:component": "BABEL_ENV=cypress cross-env FORCE_COLOR=1 cypress run --component -b chrome --spec './src/**/*.cy.ts' --project '../../../.'",
-    "test:component:open": "BABEL_ENV=cypress cross-env FORCE_COLOR=1 cypress open --component -b chrome --project '../../../.'"
-  },
-  "peerDependencies": {
-    "vue": "^3.2.47"
-  },
-  "devDependencies": {
-    "@modyfi/vite-plugin-yaml": "^1.0.4",
-    "openapi-types": "^12.1.0",
-    "vue": "^3.2.47"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/public-ui-components.git",
-    "directory": "packages/portal/spec-details"
-  },
-  "homepage": "https://github.com/Kong/public-ui-components/tree/main/packages/portal/spec-details",
-  "bugs": {
-    "url": "https://github.com/Kong/public-ui-components/issues"
-  },
-  "author": "Kong, Inc.",
-  "license": "Apache-2.0",
-  "volta": {
-    "extends": "../../../package.json"
-  },
-  "dependencies": {
-    "@kong-ui-public/swagger-ui-web-component": "workspace:^0.3.0"
-  }
-}
