--- packages/portal/spec-details/sandbox/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "baseUrl": "."
-  },
-  "include": [
-    "**/*.ts",
-    "**/*.vue",
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
