--- packages/portal/spec-details/src/index.ts
@@ -1,5 +1,0 @@
-import SpecDetails from './components/SpecDetails.vue'
-
-export default SpecDetails
-
-export * from './types'
