--- packages/portal/spec-details/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "outDir": "dist",
-    "declarationDir": "dist/types",
-    "types": [
-      "node",
-      "vite/client",
-      "cypress",
-      "../../../cypress/support"
-    ]
-  },
-  "include": [
-    "src/**/*",
-    "sandbox/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
-}
