+++ packages/portal/spec-operations-list/src/types/index.ts
@@ -1,1 +1,1 @@
-export * from './operationList'
+export * from './operation-list'
