+++ packages/portal/spec-renderer/README.md
@@ -1,5 +1,7 @@
 # @kong-ui-public/spec-renderer
 
+TODO: link to all thei individual component docs
+
 A Kong UI component for displaying and filtering API specs
 
 - [Features](#features)
