+++ packages/portal/spec-renderer/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kong-ui-public/spec-renderer",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "main": "./dist/spec-renderer.umd.js",
   "module": "./dist/spec-renderer.es.js",
@@ -42,11 +42,12 @@
   },
   "devDependencies": {
     "@modyfi/vite-plugin-yaml": "^1.0.4",
+    "openapi-types": "^12.1.0",
     "vue": "^3.2.45"
   },
   "dependencies": {
-    "@kong-ui-public/spec-details": "workspace:^0.5.0",
-    "@kong-ui-public/spec-operations-list": "^0.6.1"
+    "@kong-ui-public/spec-operations-list": "^0.6.1",
+    "@kong-ui-public/swagger-ui-web-component": "workspace:^0.3.0"
   },
   "repository": {
     "type": "git",
