+++ packages/portal/spec-renderer/sandbox/spec-details-sandbox/App.vue
