+++ packages/portal/spec-renderer/src/components/SpecDetails.vue
