+++ packages/portal/spec-renderer/src/components/SpecRenderer.vue
@@ -32,11 +32,10 @@
 
 <script setup lang="ts">
 import { PropType, ref, computed } from 'vue'
-import type { SpecDocument } from '@kong-ui-public/spec-details'
+import type { SpecDocument } from '../types'
 import type { Operation, Tag } from '@kong-ui-public/spec-operations-list'
-import SpecDetails from '@kong-ui-public/spec-details'
+import SpecDetails from './SpecDetails.vue'
 import OperationsList from '@kong-ui-public/spec-operations-list'
-import '@kong-ui-public/spec-details/dist/style.css'
 import '@kong-ui-public/spec-operations-list/dist/style.css'
 
 const props = defineProps({
