+++ packages/portal/spec-renderer/src/index.ts
@@ -1,14 +1,9 @@
-import type { App } from 'vue'
 import SpecRenderer from './components/SpecRenderer.vue'
+import SpecDetails from './components/SpecDetails.vue'
 
-// Export Vue plugin as the default
-export default {
-  // Customize Vue plugin options as desired
-  // Providing a `name` property allows for customizing the registered
-  // name of your component (useful if exporting a single component).
-  install: (app: App, options: { name?: string, [key: string]: any } = {}): void => {
-    app.component(options.name || 'SpecRenderer', SpecRenderer)
-  },
+export {
+  SpecRenderer,
+  SpecDetails,
 }
 
-export { SpecRenderer }
+export * from './types'
