+++ packages/portal/spec-renderer/src/swagger-ui-web-component.d.ts
